Rename: common package to global
