+++ src/main/java/mejai/mejaigg/global/validation/FutureDate.java
@@ -1,4 +1,4 @@
-package mejai.mejaigg.common.validation;
+package mejai.mejaigg.global.validation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
