+++ src/main/java/mejai/mejaigg/global/validation/FutureDateValidator.java
@@ -1,8 +1,8 @@
-package mejai.mejaigg.common.validation;
+package mejai.mejaigg.global.validation;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
-import mejai.mejaigg.common.util.YearMonthToEpochUtil;
+import mejai.mejaigg.global.util.YearMonthToEpochUtil;
 import mejai.mejaigg.user.dto.request.UserStreakRequest;
 
 public class FutureDateValidator implements ConstraintValidator<FutureDate, UserStreakRequest> {
