+++ src/main/java/mejai/mejaigg/user/dto/request/UserStreakRequest.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
-import mejai.mejaigg.common.validation.FutureDate;
+import mejai.mejaigg.global.validation.FutureDate;
 
 @Data
 @FutureDate
